Avoid problems on 32-bit systems (#1157)

+++ jwt/jwt_test.go
@@ -1117,7 +1117,7 @@ func TestGH375(t *testing.T) {
 
 type Claim struct {
 	Foo string
-	Bar int
+	Bar int64
 }
 
 func TestJWTParseWithTypedClaim(t *testing.T) {
